feat(pages): Add static tag release bagde image
The default github release tag is only avalaible if you make a release note on github for your tag. Like I'm lazy of doing it all the time I put a static badge only related to my current git (not to the github release api).
A better way to fix this is to cache the changelog add populate the pushed tag with it content (see #35).